requirements.txt
# enterprise-mcp-ai-platform/requirements.txt
# =============================================
# Core dependencies for production deployment.
# For local development + testing, also install requirements-dev.txt

# ── MCP Protocol ─────────────────────────────────────────────────────────────
mcp>=1.0.0                        # Model Context Protocol SDK

# ── PII Detection & Anonymization ────────────────────────────────────────────
presidio-analyzer>=2.2.35         # NLP-based PII entity recognition
presidio-anonymizer>=2.2.35       # PII masking, tokenization, encryption
spacy>=3.7.0                      # NLP backend for Presidio
# After install, run: python -m spacy download en_core_web_lg

# ── Vector Database ───────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9            # PostgreSQL driver
pgvector>=0.2.4                   # pgvector Python client

# ── AWS SDK ───────────────────────────────────────────────────────────────────
boto3>=1.35.0                     # AWS SDK (Bedrock, S3, Secrets Manager, Lake Formation)
botocore>=1.35.0

# ── RAG Evaluation ────────────────────────────────────────────────────────────
ragas>=0.1.9                      # RAG evaluation framework (faithfulness, relevancy, precision)
datasets>=2.18.0                  # HuggingFace datasets (required by RAGAS)

# ── LLM / Agent Orchestration ────────────────────────────────────────────────
langchain>=0.2.0                  # Agent orchestration
langchain-aws>=0.1.0              # LangChain Bedrock integration
langchain-mcp>=0.1.0              # LangChain MCP toolkit
llama-index-core>=0.10.0          # RAG pipeline orchestration
llama-index-vector-stores-postgres>=0.1.0  # pgvector integration for LlamaIndex

# ── Data Processing ───────────────────────────────────────────────────────────
pandas>=2.1.0
pyarrow>=14.0.0                   # Apache Arrow / Parquet
pyiceberg>=0.6.0                  # Apache Iceberg Python client
sqlalchemy>=2.0.0                 # ORM / database abstraction

# ── Streaming ─────────────────────────────────────────────────────────────────
confluent-kafka>=2.3.0            # Apache Kafka client

# ── Lineage ───────────────────────────────────────────────────────────────────
openlineage-python>=1.0.0         # OpenLineage client
openlineage-integration-common>=1.0.0

# ── Configuration & Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.0              # .env file loading
pyyaml>=6.0.1                     # YAML config parsing
pydantic>=2.5.0                   # Data validation
pydantic-settings>=2.1.0          # Settings management
tenacity>=8.2.3                   # Retry logic for external calls
structlog>=24.1.0                 # Structured logging

# ── Async ─────────────────────────────────────────────────────────────────────
anyio>=4.2.0
httpx>=0.27.0
